refactor(aitools): split texttools and listtools into smaller files per tool
