+++ CHANGELOG.md
@@ -26,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `evaluateList` and `filterList` AI tools to `list_evaluate` and `list_filter`
 - Renamed `evaluateText` and `generateText` AI tools to `text_evaluate` and `text_generate`
 - Migrated `GhPutTools` to `Utils` in `Core.Grasshopper`
+- Split AI Tools into smaller files:
+  - `TextTools` into `text_evaluate.cs` and `text_generate.cs`
+  - `ListTools` into `list_evaluate.cs` and `list_filter.cs`
 - Now `Put` removes all default inputs and outputs from the component before adding a new script component.
 - Improved OpenAI provider to support structured output.
 - Improved `script_new` in several ways:
