+++ src/SmartHopper.Core.Grasshopper/Tools/text_generate.cs
@@ -25,29 +25,14 @@ namespace SmartHopper.Core.Grasshopper.Tools
     /// <summary>
     /// Contains tools for text analysis and manipulation using AI.
     /// </summary>
-    public class TextTools : IAIToolProvider
+    public class text_generate : IAIToolProvider
     {
-        #region IAIToolProvider Implementation
-
         /// <summary>
         /// Get all tools provided by this class.
         /// </summary>
         /// <returns>Collection of AI tools.</returns>
         public IEnumerable<AITool> GetTools()
         {
-            yield return new AITool(
-                name: "text_evaluate",
-                description: "Evaluates a text against a true/false question",
-                parametersSchema: @"{
-                    ""type"": ""object"",
-                    ""properties"": {
-                        ""text"": { ""type"": ""string"", ""description"": ""The text to evaluate"" },
-                        ""question"": { ""type"": ""string"", ""description"": ""The true/false question to evaluate"" }
-                    },
-                    ""required"": [""text"", ""question"" ]
-                }",
-                execute: this.EvaluateTextToolWrapper);
-
             yield return new AITool(
                 name: "text_generate",
                 description: "Generates text based on a prompt and optional instructions",
@@ -66,57 +51,6 @@ namespace SmartHopper.Core.Grasshopper.Tools
                     ""required"": [""prompt""]
                 }",
                 execute: this.GenerateTextToolWrapper);
-        }
-
-        /// <summary>
-        /// Tool wrapper for the EvaluateText function.
-        /// </summary>
-        /// <param name="parameters">Parameters passed from the AI.</param>
-        /// <returns>Result object.</returns>
-        private async Task<object> EvaluateTextToolWrapper(JObject parameters)
-        {
-            try
-            {
-                Debug.WriteLine("[TextTools] Running EvaluateTextToolWrapper");
-
-                // Extract parameters
-                string providerName = parameters["provider"]?.ToString() ?? string.Empty;
-                string modelName = parameters["model"]?.ToString() ?? string.Empty;
-                string? text = parameters["text"]?.ToString();
-                string? question = parameters["question"]?.ToString();
-
-                if (string.IsNullOrEmpty(text) || string.IsNullOrEmpty(question))
-                {
-                    return new
-                    {
-                        success = false,
-                        error = "Missing required parameters",
-                    };
-                }
-
-                // Execute the tool
-                var result = await EvaluateTextAsync(
-                    new GH_String(text),
-                    new GH_String(question),
-                    messages => AIUtils.GetResponse(providerName, modelName, messages)).ConfigureAwait(false);
-
-                // Return standardized result
-                return new
-                {
-                    success = result.Success,
-                    result = result.Success ? result.Result.Value : false,
-                    error = result.Success ? null : result.ErrorMessage,
-                };
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[TextTools] Error in EvaluateTextToolWrapper: {ex.Message}");
-                return new
-                {
-                    success = false,
-                    error = $"Error: {ex.Message}",
-                };
-            }
         }
 
         /// <summary>
@@ -169,99 +103,6 @@ namespace SmartHopper.Core.Grasshopper.Tools
                 };
             }
         }
-
-        #endregion
-
-        #region Text Evaluation
-
-        /// <summary>
-        /// Evaluates text against a true/false question using AI with a custom GetResponse function.
-        /// </summary>
-        /// <param name="text">The text to analyze.</param>
-        /// <param name="question">The true/false question to evaluate.</param>
-        /// <param name="getResponse">Custom function to get AI response.</param>
-        /// <returns>Evaluation result containing the AI response, parsed result, and any error information.</returns>
-        public static async Task<AIEvaluationResult<GH_Boolean>> EvaluateTextAsync(
-            GH_String text,
-            GH_String question,
-            Func<List<KeyValuePair<string, string>>, Task<AIResponse>> getResponse)
-        {
-            try
-            {
-                // Prepare messages for the AI
-                var messages = new List<KeyValuePair<string, string>>
-                {
-                    // System prompt
-                    new ("system",
-                        "You are a text evaluator. Your task is to analyze a text and return a boolean value indicating whether the text matches the given criteria.\n\n" +
-                        "Respond with TRUE or FALSE, nothing else.\n\n" +
-                        "In case the text does not match the criteria, respond with FALSE."),
-
-                    // User message
-                    new ("user",
-                        $"This is my question: \"{question.Value}\"\n\n" +
-                        $"Answer the previous question based on the following input:\n{text.Value}\n\n"),
-                };
-
-                // Get response using the provided function
-                var response = await getResponse(messages).ConfigureAwait(false);
-
-                // Check for API errors
-                if (response.FinishReason == "error")
-                {
-                    return AIEvaluationResult<GH_Boolean>.CreateError(
-                        response.Response,
-                        GH_RuntimeMessageLevel.Error,
-                        response);
-                }
-
-                // Parse the response
-                var parsedResult = ParsingTools.ParseBooleanFromResponse(response.Response);
-                if (parsedResult == null)
-                {
-                    return AIEvaluationResult<GH_Boolean>.CreateError(
-                        $"The AI returned an invalid response:\n{response.Response}",
-                        GH_RuntimeMessageLevel.Error,
-                        response);
-                }
-
-                // Success case
-                return AIEvaluationResult<GH_Boolean>.CreateSuccess(
-                    response,
-                    new GH_Boolean(parsedResult.Value));
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[TextTools] Error in EvaluateTextAsync: {ex.Message}");
-                return AIEvaluationResult<GH_Boolean>.CreateError(
-                    $"Error evaluating text: {ex.Message}",
-                    GH_RuntimeMessageLevel.Error);
-            }
-        }
-
-        /// <summary>
-        /// Evaluates text against a true/false question using AI with the default AIUtils.GetResponse.
-        /// </summary>
-        /// <param name="text">The text to analyze.</param>
-        /// <param name="question">The true/false question to evaluate.</param>
-        /// <param name="provider">The AI provider to use.</param>
-        /// <param name="model">The model to use, or empty for default.</param>
-        /// <returns>Evaluation result containing the AI response, parsed result, and any error information.</returns>
-        public static Task<AIEvaluationResult<GH_Boolean>> EvaluateTextAsync(
-            GH_String text,
-            GH_String question,
-            string provider,
-            string model = "")
-        {
-            return EvaluateTextAsync(
-                text,
-                question,
-                messages => AIUtils.GetResponse(provider, model, messages));
-        }
-
-        #endregion
-
-        #region Text Generation
 
         /// <summary>
         /// Generates text from a prompt and optional instructions using AI with a custom GetResponse function.
@@ -335,7 +176,5 @@ namespace SmartHopper.Core.Grasshopper.Tools
                 instructions,
                 messages => AIUtils.GetResponse(provider, model, messages));
         }
-
-        #endregion
     }
 }
